requirements.txt
aniso8601==9.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
loginid==0.1.5
MarkupSafe==2.1.2
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.14.0
